# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PostgreSQL MCP Server",
    "version": "0.0.13",
    "description": "REST API for the PostgreSQL MCP Server. Exposes a health endpoint, MCP Streamable-HTTP transport, and an admin API for token management. Tokens and their optional per-token database connections are stored in a local JSON file (TOKENS_FILE)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local / HTTP"
    },
    {
      "url": "https://localhost:3000",
      "description": "Local / HTTPS"
    }
  ],
  "components": {
    "securitySchemes": {
      "AdminToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Value of the AUTH_TOKEN environment variable. Required for /admin/tokens and /mcp."
      },
      "FileToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "A file token created via POST /admin/tokens. Grants access to /mcp only. Each file token can optionally route to a different PostgreSQL database."
      }
    },
    "schemas": {
      "TokenConnection": {
        "type": "object",
        "nullable": true,
        "description": "Per-token PostgreSQL connection. When null the server's default connection (PG_HOST / PG_DATABASE / … env vars) is used.",
        "properties": {
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer"
          },
          "database": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "ssl": {
            "type": "string",
            "enum": [
              "false",
              "true"
            ],
            "default": "false",
            "description": "false = no TLS, true = TLS without cert verification"
          }
        }
      },
      "Token": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_used_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "connection": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TokenConnection"
              },
              {
                "type": "null"
              }
            ],
            "description": "Custom PostgreSQL connection for this token, or null to use the server default."
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/info": {
      "get": {
        "summary": "Server info",
        "description": "Returns the server name, version and the default PostgreSQL connection configuration (host, port, database, user, ssl mode). The password is never included. Requires the admin token when AUTH_TOKEN is set.",
        "operationId": "getInfo",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Server name, version and default DB connection config",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "db": {
                      "type": "object",
                      "properties": {
                        "host": {
                          "type": "string"
                        },
                        "port": {
                          "type": "integer"
                        },
                        "database": {
                          "type": "string"
                        },
                        "user": {
                          "type": "string"
                        },
                        "ssl": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "name": "pg-mcp-server",
                  "version": "0.0.13",
                  "db": {
                    "host": "localhost",
                    "port": 5432,
                    "database": "postgres",
                    "user": "postgres",
                    "ssl": "false"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Returns server status and whether TLS is enabled. No authentication required.",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Server is running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "tls": {
                      "type": "boolean"
                    }
                  }
                },
                "example": {
                  "status": "ok",
                  "tls": false
                }
              }
            }
          }
        }
      }
    },
    "/mcp": {
      "post": {
        "summary": "MCP Streamable-HTTP endpoint",
        "description": "MCP protocol transport. Requires a Bearer token (admin or file token). When the token has a custom connection configured, that PostgreSQL database is used for this session; otherwise the server default is used. Multi-step protocol: send `initialize` first, then `notifications/initialized`, then tool calls. Include the `mcp-session-id` header returned from `initialize` in subsequent requests.",
        "operationId": "mcpRequest",
        "security": [
          {
            "AdminToken": []
          },
          {
            "FileToken": []
          }
        ],
        "parameters": [
          {
            "name": "mcp-session-id",
            "in": "header",
            "required": false,
            "description": "Session ID returned by the server after initialize. Include in all follow-up requests.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jsonrpc": {
                    "type": "string"
                  },
                  "id": {
                    "type": "integer"
                  },
                  "method": {
                    "type": "string"
                  },
                  "params": {
                    "type": "object"
                  }
                }
              },
              "example": {
                "jsonrpc": "2.0",
                "id": 1,
                "method": "initialize",
                "params": {
                  "protocolVersion": "2024-11-05",
                  "capabilities": {},
                  "clientInfo": {
                    "name": "my-client",
                    "version": "1.0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MCP JSON-RPC response (may be plain JSON or text/event-stream SSE)",
            "headers": {
              "mcp-session-id": {
                "description": "Session ID — include in subsequent requests",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          },
          "403": {
            "description": "Token not authorized for this endpoint"
          },
          "503": {
            "description": "AUTH_TOKEN not configured on the server"
          }
        }
      }
    },
    "/admin/tokens": {
      "get": {
        "summary": "List all tokens",
        "description": "Returns all tokens from the token store. The token_hash field is never included in responses.",
        "operationId": "listTokens",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "All tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tokens": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Token"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token"
          },
          "503": {
            "description": "AUTH_TOKEN not configured on the server"
          }
        }
      },
      "post": {
        "summary": "Create a new token",
        "description": "Creates a new token entry in the token store. The plaintext token is returned once only — only its SHA-256 hash is persisted. An optional `connection` object routes this token to a specific PostgreSQL database; omit it (or set null) to use the server default.",
        "operationId": "createToken",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "connection": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/TokenConnection"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "description": "Optional PostgreSQL connection for this token. Omit or set null to use the server default."
                  }
                },
                "required": [
                  "name"
                ]
              },
              "example": {
                "name": "claude-desktop"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token created — plaintext value shown once",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Token"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "description": "Plaintext token — store securely, not shown again"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "name is missing or blank, or connection is not a valid object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token"
          },
          "503": {
            "description": "AUTH_TOKEN not configured on the server"
          }
        }
      }
    },
    "/admin/tokens/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "patch": {
        "summary": "Update a token",
        "description": "Updates one or more of name, active, or connection. At least one field must be provided. Set connection to null to clear a custom connection and fall back to the server default.",
        "operationId": "updateToken",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "active": {
                    "type": "boolean"
                  },
                  "connection": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/TokenConnection"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "description": "Set a custom connection, update it, or pass null to clear it."
                  }
                }
              },
              "example": {
                "name": "new-name"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated token (token_hash is never returned)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "400": {
            "description": "No valid fields provided, or connection is not a valid object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token"
          },
          "404": {
            "description": "Token not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "AUTH_TOKEN not configured on the server"
          }
        }
      },
      "delete": {
        "summary": "Delete a token",
        "description": "Permanently removes the token record from the token store. The token is immediately revoked.",
        "operationId": "deleteToken",
        "security": [
          {
            "AdminToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token"
          },
          "404": {
            "description": "Token not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "AUTH_TOKEN not configured on the server"
          }
        }
      }
    }
  }
}